✨보관함 기능 로직 추가 및 파일 분리 (#30)
