+++ Projects/Features/Scene/StorageBoxScene/StorageBox.swift
@@ -14,48 +14,48 @@ public struct StorageBox {
     @ObservableState
     public struct State: Equatable {
         public static let initialState = State()
-
+        
         public var newFolderPopupIsPresented: Bool = false
         public var editFolderPopupIsPresented: Bool = false
         public var removeFolderPopupIsPresented: Bool = false
-
+        
         public var editingIndex: Int?
-
+        
         public var defaultFolders: [StorageBoxModel] = [
             .init(title: "모든 링크", count: 3),
             .init(title: "즐겨찾기", count: 3),
             .init(title: "나중에 읽을 링크", count: 3)
         ]
-
+        
         public var customFolders: [StorageBoxModel] = [
             .init(title: "에스파", count: 3),
             .init(title: "아이브", count: 3),
             .init(title: "카리나", count: 300),
             .init(title: "A", count: 3)
         ]
-
+        
         public init() {}
     }
-
+    
     public enum Action: BindableAction {
         case onAppear
-
+        
         case storageBoxTapped(section: Int, index: Int)
-        case onEdit(index: Int? = nil)
+        case onEdit(index: Int)
         case routeToFeed(title: String)
         case tapNewFolderButton
         case addNewFolder(String)
         case removeFolder
         case cancelRemoveFolder
-
+        
         case showRemoveFolderPopup
         case routeToChangeFolderName
-
+        
         case binding(BindingAction<State>)
     }
-
+    
     public init() {}
-
+    
     public var body: some ReducerOf<Self> {
         BindingReducer()
         Reduce { state, action in
@@ -65,9 +65,7 @@ public struct StorageBox {
                 return .send(.routeToFeed(title: title))
             case let .onEdit(index):
                 print("onEdit index : \(index)")
-                if let index {
-                    state.editingIndex = index
-                }
+                state.editingIndex = index
                 state.editFolderPopupIsPresented = true
                 return .none
             case .routeToFeed:
@@ -78,6 +76,7 @@ public struct StorageBox {
                 state.newFolderPopupIsPresented = true
                 return .none
             case let .addNewFolder(folderName):
+                print("add new folder")
                 state.customFolders.append(.init(title: folderName, count: 0))
                 return .none
             case .removeFolder:
